STUD-73749 Project is closed due to internal error that is triggered when attempting to open a xaml file from Errors panel
ScriptingJitCompiler compiles expressions into a collectible AssemblyLoadContext. This ALC has a resolving event which is triggered when an assembly needed by a type is not found.
Sometimes (rarely reproducible) the expression build process triggers events in Studio that execute under the collectible alc, and when a type that does not exists is resolved, the Resolving event is triggered.
Our implementation of the Resolving event is wrong. It is searching for assemblies in our current domain and if it can't find them it calls LoaddFromAssemblyName. But LoadFromAssemblyName was already called and it triggered the Resolving event. This ends up in a loop that eventually triggers a StackOverflow exception.